requirements.txt
flask
kafka-python
confluent_kafka
